+++ src/word.h
@@ -18,13 +18,13 @@ typedef struct {
     byte   *text;
 } word_t;
 
-extern word_t  *word_new(const byte *text, size_t leng);
+extern word_t  *word_new(const byte *text, uint leng);
 extern void 	word_free(word_t *self);
 extern word_t  *word_dup(const word_t *self);
 extern word_t  *word_cpy(word_t *dst, const word_t *src);
 extern int 	word_cmp(const word_t *w1, const word_t *w2);
 extern word_t  *word_concat(const word_t *w1, const word_t *w2);
-extern void 	word_puts(const word_t *self, size_t width, FILE *fp);
+extern void 	word_puts(const word_t *self, uint width, FILE *fp);
 
 #endif	/* WORD_H */
